[WIP] Move manual numbering into own module
It isn't used by all counters
